Update enrollment service to validate user

+++ src/main/java/com/swm/studywithmentor/service/impl/EnrollmentServiceImpl.java
@@ -4,10 +4,13 @@ import com.swm.studywithmentor.model.dto.EnrollmentDto;
 import com.swm.studywithmentor.model.dto.query.SearchRequest;
 import com.swm.studywithmentor.model.dto.query.SearchSpecification;
 import com.swm.studywithmentor.model.entity.enrollment.Enrollment;
+import com.swm.studywithmentor.model.entity.user.User;
+import com.swm.studywithmentor.model.exception.ActionConflict;
 import com.swm.studywithmentor.model.exception.ApplicationException;
+import com.swm.studywithmentor.model.exception.ConflictException;
 import com.swm.studywithmentor.repository.EnrollmentRepository;
-import com.swm.studywithmentor.repository.UserRepository;
 import com.swm.studywithmentor.service.EnrollmentService;
+import com.swm.studywithmentor.service.UserService;
 import com.swm.studywithmentor.util.ApplicationMapper;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -26,12 +29,12 @@ public class EnrollmentServiceImpl implements EnrollmentService {
     private final EnrollmentRepository enrollmentRepository;
     private final ApplicationMapper applicationMapper;
     private final Function<UUID, Enrollment> findById;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
-    public EnrollmentServiceImpl(EnrollmentRepository enrollmentRepository, ApplicationMapper applicationMapper, UserRepository userRepository) {
+    public EnrollmentServiceImpl(EnrollmentRepository enrollmentRepository, ApplicationMapper applicationMapper, UserService userService) {
         this.enrollmentRepository = enrollmentRepository;
         this.applicationMapper = applicationMapper;
-        this.userRepository = userRepository;
+        this.userService = userService;
         findById = id -> enrollmentRepository.findById(id)
                 .orElseThrow(() -> new ApplicationException("NOT_FOUND", HttpStatus.NOT_FOUND , "Not found enrollment " + id.toString()));
     }
@@ -60,9 +63,7 @@ public class EnrollmentServiceImpl implements EnrollmentService {
     @Override
     public EnrollmentDto createEnrollment(EnrollmentDto enrollmentDto) {
         var enrollment = applicationMapper.enrollmentToEntity(enrollmentDto);
-        //TODO: use UserService to get user
-        var user = userRepository.findById(enrollmentDto.getStudent().getId())
-                .orElseThrow(() -> new ApplicationException("Not Found", HttpStatus.NOT_FOUND, "Not found user: " + enrollmentDto.getStudent().getId().toString()));
+        User user = userService.getCurrentUser();
         enrollment.setStudent(user);
         enrollment = enrollmentRepository.save(enrollment);
         return applicationMapper.enrollmentToDto(enrollment);
@@ -71,13 +72,10 @@ public class EnrollmentServiceImpl implements EnrollmentService {
     @Override
     public EnrollmentDto updateEnrollment(EnrollmentDto enrollmentDto) {
         var enrollment = findById.apply(enrollmentDto.getId());
-        var user = enrollmentDto.getStudent();
+        User user = userService.getCurrentUser();
         // FIXME: Optimistic locking
-        if(user != null && user.getId() != null && !enrollment.getStudent().getId().toString().equals(enrollmentDto.getStudent().getId().toString())) {
-            //TODO: use UserService to get user
-            var newUser = userRepository.findById(enrollmentDto.getStudent().getId())
-                    .orElseThrow(() -> new ApplicationException("Not Found", HttpStatus.NOT_FOUND, "Not found user: " + enrollmentDto.getStudent().getId().toString()));
-            enrollment.setStudent(newUser);
+        if(!enrollment.getStudent().getId().equals(user.getId())) {
+            throw new ConflictException(Enrollment.class, ActionConflict.UPDATE, "User does not own this enrollment", user.getId());
         }
         applicationMapper.enrollmentToEntity(enrollment, enrollmentDto);
         enrollment = enrollmentRepository.save(enrollment);
@@ -88,6 +86,10 @@ public class EnrollmentServiceImpl implements EnrollmentService {
     @Override
     public void deleteEnrollment(UUID id) {
         var enrollment = findById.apply(id);
+        User user = userService.getCurrentUser();
+        if(!enrollment.getStudent().getId().equals(user.getId())) {
+            throw new ConflictException(Enrollment.class, ActionConflict.DELETE, "User does not own this enrollment", user.getId());
+        }
         enrollmentRepository.delete(enrollment);
     }
 }
